Major services refactor

+++ src/main/java/pl/sglebocki/spring/blog/services/PostsService.java
@@ -9,10 +9,10 @@ import pl.sglebocki.spring.blog.dto.PostSaveDTO;
 public interface PostsService {
 
 	public Collection<PostShowDTO> getNewestPosts(int number);
-	public PostShowDTO getPostById(int postId);
+	public PostShowDTO getPostById(long postId);
 	public PostShowDTO getPostByIdWithAuthentication(String name, Integer postId);
 	public void saveOrUpdatePost(String userName,PostSaveDTO post);
 	public Collection<PostShowDTO> getPostsByUserName(String userName, DatePeriodDTO datePeriod);
-	public void delatePostById(String userName, int postId);
+	public void delatePostById(String userName, long postId);
 	
 }
